requirements.txt
# Core dependencies
requests>=2.31.0
PyGithub>=2.1.1
python-dotenv>=1.0.0

# Code analysis
pygments>=2.17.0
gitpython>=3.1.40

# Utilities
pyyaml>=6.0.1
jsonschema>=4.20.0
colorama>=0.4.6

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.1

# Type checking
mypy>=1.7.1
types-requests>=2.31.0
types-PyYAML>=6.0.12

# Code quality
black>=23.12.0
flake8>=6.1.0
pylint>=3.0.3
